Refactor taskCards.js: Improve rendering of category menu
The renderCardsMenuTemplate function in taskCards.js has been modified to improve the rendering of the category menu. Specifically, the code now checks if the current status is "done" before adding the "Add Task" button. This change ensures that the "Add Task" button is only displayed for categories other than "done".

Closes #123