Refactor user login and logout functions
This commit refactors the code by extracting the user login and logout functions into separate functions: isUserLoggedIn() and userLogOut(). The isUserLoggedIn() function checks if the user is logged in by retrieving the logged-in user ID from the local storage. If the user is not logged in, a console log message is displayed. The userLogOut() function removes the logged-in user ID from the local storage and redirects the user to the index.html page.